refactor(core): update project references and namespace
