+++ src/Lunar.Framework.MooaLewaUI.Test/Lunar.Framework.MooaLewaUI.Test.csproj
@@ -23,7 +23,7 @@
 
     <ItemGroup>
         <ProjectReference Include="..\Lunar.Framework.MooaLewaUI.SourceGenerator\Lunar.Framework.MooaLewaUI.SourceGenerator.csproj"/>
-        <ProjectReference Include="..\Lunar.Framework.MooaLewaUI.Sample\Lunar.Framework.MooaLewaUI.Sample.csproj" />
+        <ProjectReference Include="..\Lunar.Framework.MooaLewaUI\Lunar.Framework.MooaLewaUI.csproj" />
     </ItemGroup>
 
 </Project>
