+++ src/Lunar.Framework.MooaLewaUI/Lunar.Framework.MooaLewaUI.csproj
@@ -7,18 +7,11 @@
     </PropertyGroup>
 
     <ItemGroup>
-        <ProjectReference Include="..\Lunar.Framework.MooaLewaUI\Lunar.Framework.MooaLewaUI.csproj"/>
         <ProjectReference Include="..\Lunar.Core.Base\Lunar.Core.Base.csproj"/>
     </ItemGroup>
 
     <ItemGroup>
         <PackageReference Include="Microsoft.CodeAnalysis.CSharp" Version="3.8.0"/>
-    </ItemGroup>
-
-    <ItemGroup>
-        <ProjectReference Include="..\Lunar.Framework.MooaLewaUI.SourceGenerator\Lunar.Framework.MooaLewaUI.SourceGenerator.csproj"
-                          OutputItemType="Analyzer"
-                          ReferenceOutputAssembly="false"/>
     </ItemGroup>
 
     <ItemGroup>
